CSV Export: Fix modified price output snippet to better support number of decimal settings higher than 2
The original version of this snippet used the price values directly from the CSV which works well when the price decimals are set to 2 or less. However, when you have price decimals set to anything above, your totals will be exported with the right number of decimals but they wouldn't match the store's values since CSV Export already rounded them to 2 decimal places before this snippet attempted to format to more decimal places.

This update ensures the raw values from WooCommerce are used for all totals before rounding to the correct number of decimal places. Due to the differentces between the line item, fee, shipping, and tax filters, this snippet needed to be expanded considerably.